requirements.txt
absl-py==2.3.0
audioread==3.0.1
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
diffsptk==2.1.0
einops==0.8.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
grpcio==1.73.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.3
pysptk==1.0.1
python-dateutil==2.9.0.post0
pyworld==0.3.5
PyYAML==6.0.2
requests==2.32.4
resampy==0.4.3
scikit-learn==1.7.0
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tomli==2.2.1
torchcomp==0.2.1
torchcrepe==0.0.24
torchlpc==0.6
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
vector_quantize_pytorch==1.12.12
Werkzeug==3.1.3